requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.7.6
PyNaCl>=1.4.0
python-dotenv>=1.0.0
requests>=2.31.0
ffmpeg-python>=0.2.0
